# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ADP 鱼塘养殖企业 API",
    "version": "1.0.0",
    "description": "网站二正式业务接口。提交后可修改并递增版本，核验后只读。"
  },
  "servers": [
    {
      "url": "https://1.14.148.15",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/cost/allocation-rules": {
      "get": {
        "tags": [
          "cost"
        ],
        "summary": "查询列表",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_allocation_rules_get_api_v1_cost_allocation_rules",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "cost"
        ],
        "summary": "保存配置",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_save_allocation_rules_put_api_v1_cost_allocation_rules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
...
    "/api/v1/cost/entries": {
      "post": {
        "tags": [
          "cost"
        ],
        "summary": "新建记录",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_create_entry_post_api_v1_cost_entries",
        "parameters": [],
        "responses": {
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "cost"
        ],
        "summary": "查询列表",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_entries_get_api_v1_cost_entries",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/v1/cost/entries/{entry_id}": {
      "delete": {
        "tags": [
          "cost"
        ],
        "summary": "删除未提交草稿",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_delete_draft_delete_api_v1_cost_entries_entry_id",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      },
      "patch": {
        "tags": [
          "cost"
        ],
        "summary": "更新",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_update_entry_patch_api_v1_cost_entries_entry_id",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/cost/entries/{entry_id}/confirm": {
      "post": {
        "tags": [
          "cost"
        ],
        "summary": "确认入账",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_confirm_entry_post_api_v1_cost_entries_entry_id_confirm",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/cost/entries/{entry_id}/reverse": {
      "post": {
        "tags": [
          "cost"
        ],
        "summary": "冲销",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_reverse_entry_post_api_v1_cost_entries_entry_id_reverse",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/cost/entries/{entry_id}/submit": {
      "post": {
        "tags": [
          "cost"
        ],
        "summary": "提交核验",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_submit_entry_post_api_v1_cost_entries_entry_id_submit",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/cost/entries/{entry_id}/verify": {
      "post": {
        "tags": [
          "cost"
        ],
        "summary": "核验",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_verify_entry_post_api_v1_cost_entries_entry_id_verify",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
...
    "/api/v1/cost/structure": {
      "get": {
        "tags": [
          "cost"
        ],
        "summary": "查询列表",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "cost_structure_get_api_v1_cost_structure",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "adp_session"
      },
      "CsrfSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          },
          "data": {},
          "request_id": {
            "type": "string",
            "example": "request-example-001"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "VERSION_CONFLICT"
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          },
          "data": {},
          "request_id": {
            "type": "string",
            "example": "request-example-001"
          }
        }
      },
      "MutationRequest": {
        "type": "object",
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "evidence_attachment_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          }
        },
        "additionalProperties": true
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "统一错误；以 code 和 request_id 定位问题",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}